SearchProcessor interface, data and pep objects, Integer and Numeric parameters and ProcesingHistory object

+++ src/main/java/br/com/isaquebrb/iftm/batchcreditanalysis/model/dto/ParameterResponse.java
@@ -3,23 +3,18 @@ package br.com.isaquebrb.iftm.batchcreditanalysis.model.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Builder
 @Getter
 public class ParameterResponse {
 
     private Long id;
-
     private String name;
-
     private String description;
-
     private String stringValue;
-
     private Integer integerValue;
-
-    private Double numericValue;
-
+    private BigDecimal numericValue;
     private Boolean booleanValue;
-
     private Boolean active;
 }
